Csl::Citation - extract class Csl::Mapper
